@@ -9,7 +9,7 @@ readme = "README.md"
 packages = [{include = "async_signals"}]
 
 [tool.poetry.dependencies]
-python = "^3.8"
+python = "^3.9"
 
 [tool.poetry.group.dev.dependencies]
 pytest = ">=7.1.3,<9.0.0"
@@ -22,7 +22,7 @@ pyright = ">=1.1.350,<1.2"
 
 [tool.ruff]
 line-length = 115
-target-version = "py38"
+target-version = "py39"
 output-format = "grouped"
 
 [tool.ruff.lint]
