@@ -1,6 +1,6 @@
 import functools
 import inspect
-from typing import Callable, Tuple
+from typing import Callable
 
 # Those functions are copied from
 # https://github.com/django/django/blob/main/django/utils/inspect.py
@@ -10,7 +10,7 @@ from typing import Callable, Tuple
 def _get_func_parameters(
     func: Callable,
     remove_first: bool,
-) -> Tuple[inspect.Parameter, ...]:
+) -> tuple[inspect.Parameter, ...]:
     parameters = tuple(inspect.signature(func).parameters.values())
     if remove_first:
         parameters = parameters[1:]
@@ -19,7 +19,7 @@ def _get_func_parameters(
 
 def _get_callable_parameters(
     meth_or_func: Callable,
-) -> Tuple[inspect.Parameter, ...]:
+) -> tuple[inspect.Parameter, ...]:
     is_method = inspect.ismethod(meth_or_func)
     func = meth_or_func.__func__ if is_method else meth_or_func  # type: ignore
     return _get_func_parameters(func, remove_first=is_method)
